Some prep work for implement Cng symmetric algorithms.
This is just code movement and renaming in anticipation
of upcoming Cng work for the retro symmetric algorithms.

We split CngAsymmetricAlgorithmCore into two pieces:

- CngKey management (applicable to both both symmetric
  and asymmetric algorithms)

- Static helpers for hashing, signing and verifying
  (asymmetric only and uses no instance data.)
